Добавили стуркутру llm-service
