+++ llm-service/redis_service.py
@@ -1,0 +1,9 @@
+import redis
+from config import REDIS_HOST, REDIS_PORT
+
+redis_client = redis.Redis(
+    host=REDIS_HOST,
+    port=REDIS_PORT,
+    db=0,
+    decode_responses=False
+)
