--- llm-service/replace.txt
@@ -1,1 +1,0 @@
-YANDEX_API_KEY_PLACEHOLDER==SECRET==
